refactor: split to AddTabFrom as component
